+++ geotrek/tourism/admin.py
@@ -47,7 +47,7 @@ class TouristicContentType2Inline(TranslationTabularInline):
 
 
 class TouristicContentCategoryAdmin(TranslationAdmin):
-    list_display = ('label', 'pictogram_img', 'type1_label', 'type2_label')
+    list_display = ('label', 'order', 'pictogram_img', 'type1_label', 'type2_label')
     search_fields = ('label',)
     inlines = [
         TouristicContentType1Inline,